% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{bp_launch_shinystan}
\alias{bp_launch_shinystan}
\title{Launch shinystan}
\usage{
bp_launch_shinystan(fit)
}
\arguments{
\item{fit}{List of output as returned by \code{stan_fit()}.}
}
\description{
Launch shinystan an interactive tool for stan model evaluation
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
bp_launch_shinystan(fit)
}
}
