% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check a data.frame}
\usage{
check_dataframe(dataframe, req_vars, req_types)
}
\arguments{
\item{dataframe}{A data.frame to check.}

\item{req_vars}{A character vector of variables that are required.}

\item{req_types}{A character vector of types for each required variable.}
}
\description{
Check a data.frame
}
\keyword{internal}
