% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mean_sd}
\alias{check_mean_sd}
\title{Check a parameter is a length is 2 numeric vector}
\usage{
check_mean_sd(param, name = "r_init")
}
\arguments{
\item{param}{A parameter to check the format of.}

\item{name}{A character string naming the variable
to check.}
}
\description{
Check a parameter is a length 2 numeric vector
defining a mean and a standard deviation.
}
