% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_observations}
\alias{check_observations}
\title{Check observations are in the correct format}
\usage{
check_observations(obs)
}
\arguments{
\item{obs}{A \code{data.frame} of observations to check for formatting issues.}
}
\description{
Check observations are in the correct format
}
\examples{
obs <- latest_obs(germany_covid19_delta_obs)
check_observations(obs)
}
\seealso{
Functions used for checking inputs∂.
\code{\link{check_dataframe}()},
\code{\link{check_param}()},
\code{\link{check_quantiles}()}
}
\concept{check}
