% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{combine_posteriors_dt}
\alias{combine_posteriors_dt}
\title{FUNCTION_TITLE}
\usage{
combine_posteriors_dt(forecasts, target = "forecast")
}
\arguments{
\item{forecasts}{A data frame of forecasts as produced by \code{forecast_dt()}}

\item{target}{A character string indicating the list of posteriors to
combine. This can be either "posterior", or "forecast".}
}
\value{
A data frame of summmarised posteriors along with summary
information about the forecast method used.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_obs), max_treedepth = 15)
dt <- combine_posteriors_dt(dt, target = "forecast")
}
}
