% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{convert_to_stanfit}
\alias{convert_to_stanfit}
\title{Convert to stanfit object}
\usage{
convert_to_stanfit(fit)
}
\arguments{
\item{fit}{List of output as returned by \code{stan_fit()}.}
}
\value{
The model fit as a stanfit object
}
\description{
Convert to stanfit object
}
\examples{
\dontrun{
obs <- latest_obs(germany_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
convert_to_stanfit(fit)
}
}
