% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract posterior draws}
\usage{
extract_draws(fit, ...)
}
\arguments{
\item{...}{Additional parameters passed to \code{cmdstanr::draws}}
}
\value{
A \code{draws} object from the \code{posterior} package.
}
\description{
Extract posterior draws
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
extract_draws(fit)
}
}
