% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_forecast_by_type}
\alias{extract_forecast_by_type}
\title{Extract forecasts from a posterior data frame by type}
\usage{
extract_forecast_by_type(posterior, forecast_dates)
}
\arguments{
\item{posterior}{A data frame of summarised posterior estimates
as returned by \code{cmdstanr::summary} with  an additional type variable
which contains the following character string options: "non-DELTA",
"DELTA", "Combined", "Overall".}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast. Must contain a "Cases" date and a "Sequence" date.
Default is to infer these dates from the summarised posterior.}
}
\value{
A data frame of forecasts from the posterior data frame.
}
\description{
Extract forecasts from a posterior data frame by type
}
