% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_forecast_dates}
\alias{extract_forecast_dates}
\title{Extract forecast dates}
\usage{
extract_forecast_dates(posterior, forecast_dates = NULL)
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast vs an estimate. Defaults to empty in which case
forecast dates are inferred from the \code{posterior} list based on data
availability for cases and sequences. These dates can be overridden
by supplying a replacement data with a duplicate name (see the examples).}
}
\value{
A named vector of dates.
}
\description{
Extract forecast dates based on the availability of both case
and sequence data. Custom forecasts dates can also be defined and the
automated forecasts can be overridden as desired.
}
\examples{
\dontrun{
options(mc.cores = 4)
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs, overdispersion = FALSE)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, max_treedepth = 15, adapt_delta = 0.9)
p <- summarise_posterior(fit)
# default
extract_forecast_dates(p)

# add a custom date
extract_forecast_dates(p, c("custom" = "2021-08-01"))

# overwrite a date
extract_forecast_dates(p, c("Cases" = "2021-08-01"))
}
}
