% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_by_availability}
\alias{filter_by_availability}
\title{Filter data based on availability and forecast date}
\usage{
filter_by_availability(
  obs,
  date = max(obs$date),
  seq_date = date,
  case_date = date
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{date}{Date at which to filter. Defaults to the
maximum date in \code{obs}.}

\item{seq_date}{Date from which to use available sequence data. Defaults to
the forecast date.}

\item{case_date}{Date from which to use available case data. Defaults to
the forecast date.}
}
\description{
Filter data based on availability and forecast date
}
\examples{
dt <- rbind(
  update_obs_availability(germany_covid19_delta_obs, seq_lag = 3),
  update_obs_availability(germany_covid19_delta_obs, seq_lag = 1)
)
# filter out duplicates and up to the present date
filter_by_availability(dt)

# filter to only use sequence data up the the 12th of June
filter_by_availability(dt, seq_date = "2021-06-12")

# as above but only use
filter_by_availability(dt,
  seq_date = "2021-06-12",
  case_date = "2021-07-01"
)
}
