% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast using branching processes at a target date}
\usage{
forecast(
  obs,
  plot_obs = bp.delta::latest_obs(obs),
  forecast_date = max(obs$date),
  seq_date = forecast_date,
  case_date = forecast_date,
  save_path = tempdir(),
  horizon = 4,
  delta = c(0.2, 0.2),
  strains = 2,
  variant_relationship = "pooled",
  overdispersion = TRUE,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  plot = TRUE,
  ...
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_date}{Date at which to forecast. Defaults to the
maximum date in \code{obs}.}

\item{seq_date}{Date from which to use available sequence data. Defaults to
the forecast date.}

\item{case_date}{Date from which to use available case data. Defaults to
the forecast date.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{delta}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model.extract}

\item{plot}{Logical, should posterior plots be produced and saved.
Defaults to \code{TRUE}.}
}
\description{
Forecast using branching processes at a target date
}
