% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.vocs-package.R
\docType{package}
\name{forecast.vocs-package}
\alias{forecast.vocs}
\alias{forecast.vocs-package}
\title{forecast.vocs: Forecast Case Notifications using Variant of Concern Strain Dynamics}
\description{
Forecast case notifications using variant of concern strain dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/forecast.vocs/}
  \item \url{https://github.com/epiforecasts/forecast.vocs/}
  \item Report bugs at \url{https://github.com/epiforecasts/forecast.vocs/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

}
\keyword{internal}
