% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_accross_dates}
\alias{forecast_accross_dates}
\title{Forecast across multiple dates}
\usage{
forecast_accross_dates(obs, forecast_dates = unique(obs)$date[-c(1:3)], ...)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_dates}{A list of dates to forecast at.}

\item{...}{Additional parameters passed to \code{forecast()}}
}
\value{
A list each containing the output from running \code{forecast()}
on a single forecast date
}
\description{
Forecast across multiple dates
}
