% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_accross_scenarios}
\alias{forecast_accross_scenarios}
\title{Forecast across multiple scenarios and dates}
\usage{
forecast_accross_scenarios(obs, scenarios, save_path = tempdir(), ...)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{scenarios}{A dataframe of scenarios as produced by
\code{define_scenarios()}. If missing uses the default scenarios
from \code{default_scenarios()}.}

\item{save_path}{Character string indicating the path to save results
to. Defaults to \code{tempdir()}. Each scenario will be saved in a sub folder.}

\item{...}{Additional parameters passed to \code{forecast_across_dates()}}
}
\value{
A list each containing the output from running
\code{forecast_across_dates()} on a single scenario.
}
\description{
Forecast across multiple scenarios and dates
}
