% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-tools.R
\name{forecast_across_scenarios}
\alias{forecast_across_scenarios}
\title{Forecast across multiple scenarios and dates}
\usage{
forecast_across_scenarios(obs, scenarios, ...)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_voc, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{scenarios}{A dataframe of scenarios as produced by
\code{define_scenarios()}. If missing uses the default scenarios
from \code{default_scenarios()}.}

\item{...}{Additional parameters passed to \code{forecast_across_dates()}.}
}
\value{
A data table each rows containing the output from running
\code{forecast()} on a single scenario for a single forecast date.
}
\description{
Forecast across multiple scenarios and dates
}
\seealso{
Functions used for forecastin across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
