% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_n_strain}
\alias{forecast_n_strain}
\title{Forecast for a single model and summarise}
\usage{
forecast_n_strain(
  data,
  model = NULL,
  strains = 2,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{stan_data()}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{load_model()}}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for.}

\item{...}{Additional parameters passed to \code{stan_fit()}.}
}
\description{
Forecast for a single model and summarise
}
