% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fv_as_data_list}
\alias{fv_as_data_list}
\title{Format data for use with stan}
\usage{
fv_as_data_list(
  obs,
  horizon = 4,
  r_init = c(0, 0.25),
  r_step = 1,
  r_forecast = TRUE,
  beta = c(0, 0.5),
  lkj = 0.5,
  voc_scale = c(0, 0.2),
  variant_relationship = "correlated",
  overdispersion = TRUE,
  likelihood = TRUE,
  output_loglik = TRUE,
  debug = FALSE
)
}
\arguments{
\item{obs}{A data frame with the following variables:
\code{date}, \code{cases}, \code{seq_voc}, and \code{seq_total}.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{r_init}{Numeric vector of length 2. Mean and
standard deviation for the normal prior on the initial log growth rate.}

\item{r_step}{Integer, defaults to 1. The number of observations between
each change in the growth rate.}

\item{r_forecast}{Logical, defaults \code{TRUE}. Should the growth rate be
forecast beyond the data horizon.}

\item{beta}{Numeric vector, defaults to c(0, 0.5). Represents the mean and
standard deviation of the normal prior (truncated at 1 and -1) on the
weighting in the differenced AR process of  the previous difference.
Placing a tight prior around zero effectively reduces the AR process to a
random walk on the growth rate.}

\item{lkj}{Numeric defaults to 0.5. The assumed prior covariance between
variants growth rates when using the "correlated" model. This sets the shape
parameter for the Lewandowski-Kurowicka-Joe (LKJ) prior distribution. If set
to 1 assigns a uniform prior for all correlations, values less than 1
indicate increased belief in strong correlations and values greater than 1
indicate increased belief weaker correlations. Our default setting places
increased weight on some correlation between strains.}

\item{voc_scale}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant of concern.}

\item{variant_relationship}{Character string, defaulting to "correlated".
Controls the relationship of strains with options being "correlated"
(strains growth rates are correlated over time), "scaled" (a fixed scaling
between strains), and "independent" (fully independent strains after
initial scaling).}

\item{overdispersion}{Logical, defaults to \code{TRUE}. Should the observations
used include overdispersion.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}
}
\value{
A list as required by stan.
}
\description{
Format data for use with stan
}
\examples{
fv_as_data_list(latest_obs(germany_covid19_delta_obs))
}
\seealso{
Functions used for modelling
\code{\link{fv_inits}()},
\code{\link{fv_model}()},
\code{\link{fv_sample}()}
}
\concept{model}
