% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fv_dow_period}
\alias{fv_dow_period}
\title{Calculate the day of the week periodicity}
\usage{
fv_dow_period(t, start_date, specials = c(), special_to = "Sunday")
}
\arguments{
\item{t}{An integer indicating the number of dates}

\item{start_date}{A date indicating the start date}

\item{specials}{A vector of special dates to modify the day of the week for.}

\item{special_to}{A character string indicating which day of the week or
other label to assign holidays. By default this is set to "Sunday"}
}
\value{
A vector indicating the period of the dates.
}
\description{
This helper function allows the user to generate a vector of day of the
the week periods.
}
\examples{
fv_dow_period(t = 10, start_date = as.Date("2021-12-01"))
}
\seealso{
Preprocessing functions
\code{\link{filter_by_availability}()},
\code{\link{latest_obs}()},
\code{\link{piecewise_steps}()}
}
\concept{preprocess}
