% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{fv_extract_forecast}
\alias{fv_extract_forecast}
\title{Extract forecasts from a summarised posterior}
\usage{
fv_extract_forecast(posterior)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}
}
\value{
A \code{data.frame} of forecasts in the format returned
by \code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}} but with fitting variables dropped.
}
\description{
Uses the \code{observed} variable returned by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}} to return posterior predictions
for forecast dates only.
}
\examples{
p <- fv_example(strains = 2, type = "posterior")

fv_extract_forecast(p)
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
