% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fv_inits}
\alias{fv_inits}
\title{Set up initial conditions for model}
\usage{
fv_inits(data, strains = 2)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=fv_data]{fv_data()}}.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}
}
\value{
A function that when called returns a list of initial conditions
for the package stan models.
}
\description{
Set up initial conditions for model
}
\examples{
dt <- fv_data(latest_obs(germany_covid19_delta_obs))
inits <- fv_inits(dt)
inits
inits()
}
\seealso{
Functions used for modelling
\code{\link{fv_data}()},
\code{\link{fv_model}()},
\code{\link{fv_sample}()}
}
\concept{model}
