% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fv_model}
\alias{fv_model}
\title{Load and compile a strain model}
\usage{
fv_model(model, include, strains = 2, compile = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{model}{A character string indicating the path to the model.
If not supplied the package default model is used.}

\item{include}{A character string specifying the path to any stan
files to include in the model. If missing the package default is used.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{compile}{Logical, defaults to \code{TRUE}. Should the model
be loaded and compiled using \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose
messages be shown.}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A \code{cmdstanr} model.
}
\description{
Load and compile a strain model
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# one strain model
mod <- fv_model(strains = 1)

# two strain model
two_strain_mod <- fv_model(strains = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for modelling
\code{\link{fv_as_data_list}()},
\code{\link{fv_inits}()},
\code{\link{fv_sample}()}
}
\concept{model}
