% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{fv_posterior}
\alias{fv_posterior}
\title{Summarise the posterior}
\usage{
fv_posterior(fit, probs = c(0.05, 0.2, 0.8, 0.95), ...)
}
\arguments{
\item{fit}{List of output as returned by \code{\link[=fv_sample]{fv_sample()}}.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{\link[=plot_cases]{plot_cases()}}, \code{\link[=plot_rt]{plot_rt()}},
and \code{\link[=plot_voc]{plot_voc()}}).}

\item{...}{Additional arguments that may be passed but will not be used.}
}
\value{
A dataframe summarising the model posterior.
}
\description{
A generic wrapper around \code{\link[posterior:draws_summary]{posterior::summarise_draws()}} with
opinionated defaults. See \code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}} for a more
opinionated wrapper with further cleaning and tidying
including linking to observed data, tidying parameter names,
and transforming parameters for interpretability.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- fv_as_data_list(obs)
inits <- fv_inits(dt)
fit <- fv_sample(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
fv_posterior(fit)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
