% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fv_sample}
\alias{fv_sample}
\title{Fit a brancing process strain model}
\usage{
fv_sample(
  data,
  model = forecast.vocs::fv_model(strains = 2),
  diagnostics = TRUE,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=fv_data]{fv_data()}}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=fv_model]{fv_model()}}.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a \code{data.frame}.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A \code{data.frame} containing the \code{cmdstanr} fit, the input data, the
fitting arguments, and optionally summary diagnostics.
}
\description{
Fit a brancing process strain model
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)

# format example data
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- fv_data(obs)

# single strain model
inits <- fv_inits(dt, strains = 1)
mod <- fv_model(strains = 1)
fit <- fv_sample(
  dt,
  model = mod, init = inits,
  adapt_delta = 0.99, max_treedepth = 15
)
fit

# two strain model
inits <- fv_inits(dt, strains = 2)

mod <- fv_model(strains = 2)

two_strain_fit <- fv_sample(dt,
  model = mod, init = inits,
  adapt_delta = 0.99, max_treedepth = 15
)
two_strain_fit
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for modelling
\code{\link{fv_data}()},
\code{\link{fv_inits}()},
\code{\link{fv_model}()}
}
\concept{model}
