% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{fv_tidy_posterior}
\alias{fv_tidy_posterior}
\title{Summarise the posterior tidily}
\usage{
fv_tidy_posterior(
  fit,
  probs = c(0.05, 0.2, 0.8, 0.95),
  digits = 3,
  voc_label = "VOC",
  scale_r = 1
)
}
\arguments{
\item{fit}{List of output as returned by \code{\link[=fv_sample]{fv_sample()}}.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{\link[=plot_cases]{plot_cases()}}, \code{\link[=plot_rt]{plot_rt()}},
and \code{\link[=plot_voc_frac]{plot_voc_frac()}}).}

\item{digits}{Numeric, defaults to 3. Number of digits to round summary
statistics to.}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{scale_r}{Numeric, defaults to 1. Rescale the timespan over which
the growth rate and reproduction number is calculated. An example use case
is rescaling the growth rate from weekly to be scaled by the mean of
the generation time (for COVID-19 for example this would be 5.5 / 7.}
}
\value{
A dataframe summarising the model posterior. Output is stratified
by \code{value_type} with posterior summaries by case, voc, voc advantage vs
non-voc over time, rt, growth, model, and the raw posterior summary.
}
\description{
A very opinionated wrapper around
\code{\link[posterior:draws_summary]{posterior::summarise_draws()}} with cleaning and tidying including linking
to observed data, tidying parameter names, and transforming parameters for
interpretability. See \code{\link[=fv_posterior]{fv_posterior()}} for a more generic solution.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- fv_as_data_list(obs)
inits <- fv_inits(dt)
fit <- fv_sample(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
fv_tidy_posterior(fit)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{plot.fv_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
