% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-availability.R
\name{generate_obs_scenario}
\alias{generate_obs_scenario}
\title{Define observed data for a scenario}
\usage{
generate_obs_scenario(obs, seq_lag, seq_samples)
}
\arguments{
\item{obs}{A dataframe of observations as returned by \code{latest_obs}
or similar.}

\item{seq_lag}{Number, number of weeks to lag sequence data behind
date of observation.}

\item{seq_samples}{Fraction of sequence samples to include.}
}
\value{
A data frame of scenario definitions with ids
}
\description{
Define observed data for a scenario
}
\examples{
generate_obs_scenario(latest_obs(germany_covid19_delta_obs), 4, 0.1)
}
