% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{germany_obs}
\alias{germany_obs}
\title{Test positive COVID-19 cases and sequences in Germany}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19 rows and 9 columns.
}
\usage{
germany_obs
}
\value{
A \code{data.table} with the following variables: date, location,
location_name, cases, seq_total, seq_delta, share_delta, cases_available,
and seq_available.
}
\description{
Test positive COVID-19 cases and sequences stratified by
DELTA variant summarised by week for Germany. Data is sourced from the RKI
via the Germany/Poland forecasting hub.
}
\keyword{datasets}
