% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{link_dates_with_posterior}
\alias{link_dates_with_posterior}
\title{Link dates by time for posterior parameter estimates}
\usage{
link_dates_with_posterior(posterior, data, mod_end = 0)
}
\arguments{
\item{posterior}{A data frame of summarised posterior estimates
as returned by \code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} with  an additional type variable
which contains the following character string options: "non-VOC",
"VOC", "Combined", "Overall".}

\item{data}{A list of data as returned in the "data" entry of the output
returned by \code{\link[=fv_sample]{fv_sample()}}.}

\item{mod_end}{Integer, defaults to 0. Amount to shift the end date of
estimates.}
}
\value{
A posterior \code{data.frame} with an additional data column.
}
\description{
Link dates by time for posterior parameter estimates
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{plot.fv_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
