% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{link_obs_with_posterior}
\alias{link_obs_with_posterior}
\title{Link posterior estimates with observed data}
\usage{
link_obs_with_posterior(posterior, obs, horizon, target_types)
}
\arguments{
\item{posterior}{A data frame of summarised posterior estimates
as returned by \code{cmdstanr::summary} with  an additional type variable
which contains the following character string options: "non-VOC",
"VOC", "Combined", "Overall".}

\item{obs}{Numeric vector of observed data to link to posterior estimates.}

\item{horizon}{Integer indicating the horizon of unobserved forecasts. If not
specified will be inferred from \code{obs}.}

\item{target_types}{A character vector of types (as specified in the \code{type}
variable) to modify.}
}
\value{
The input data.frame combined with \code{obs} and \code{observed} variables.
}
\description{
Link posterior estimates with observed data and flag if values are observed
or not.
}
