% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example}
\alias{load_example}
\title{Load a package example}
\usage{
load_example(strains = 1, type = "posterior")
}
\arguments{
\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{type}{A character string indicating the example to load.
Supported options are "posterior", "forecast", "observations",
and "script" which are the
output of \code{\link[=summarise_posterior]{summarise_posterior()}},  \code{\link[=extract_forecast]{extract_forecast()}},
\code{filter_by_availability} (with the date argument set to "2021-08-26"
applied to the \link{germany_covid19_delta_obs} package dataset),
and the script used to generate these examples respectively.}
}
\value{
A \code{data.table} of summarised output
}
\description{
Loads examples of posterior and forecast summaries produced
using example scripts. Used to streamline examples,
in package tests and to enable users to explore package functionality
without needing to install \code{cmdstanr}.
}
\examples{
# Load the summarised posterior from an example fit of the one strain model
load_example(strains = 1, type = "posterior")

# Load the summarised forecast from this posterior
load_example(strains = 1, type = "forecast")

# Load the script used to generate these examples
# Optionally source this script to regenerate the example
readLines(load_example(strains = 1, type = "script"))
}
\seealso{
Package data sets
\code{\link{germany_covid19_delta_obs}}
}
\concept{data}
