% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_model}
\alias{load_model}
\title{Load and compile a strain model}
\usage{
load_model(strains = 2, compile = TRUE, ...)
}
\arguments{
\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{compile}{Logical, defaults to \code{TRUE}. Should the model
be loaded and compiled using \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A \code{cmdstanr} model.
}
\description{
Load and compile a strain model
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# one strain model
mod <- load_model(strains = 1)

# two strain model
two_strain_mod <- load_model(strains = 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for modelling
\code{\link{stan_data}()},
\code{\link{stan_fit}()},
\code{\link{stan_inits}()}
}
\concept{model}
