% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-methods.R
\name{plot.fv_forecast}
\alias{plot.fv_forecast}
\title{Plot method for forecast}
\usage{
\method{plot}{fv_forecast}(x, target = "posterior", type, ...)
}
\arguments{
\item{x}{A \code{data.table} of output as produced by \code{\link[=forecast]{forecast()}}.}

\item{type}{A character string indicating the type of plot required.}

\item{...}{Pass additional arguments to lower level plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "fv_forecast".
}
\seealso{
Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast_n_strain}()},
\code{\link{forecast}()},
\code{\link{summary.fv_forecast}()},
\code{\link{unnest_posterior}()}

Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()}
}
\concept{forecast}
\concept{plot}
