% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{plot.fv_posterior}
\alias{plot.fv_posterior}
\title{Plot method for forecast}
\usage{
\method{plot}{fv_posterior}(
  x,
  type = "cases",
  obs = NULL,
  forecast_dates = NULL,
  all_obs = FALSE,
  voc_label = "variant of concern",
  ...
)
}
\arguments{
\item{x}{A \code{data.table} of output as produced by \code{\link[=forecast]{forecast()}}.}

\item{type}{A character string indicating the type of plot required.}

\item{...}{Pass additional arguments to lower level plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "fv_forecast".
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")

}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}

Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()},
\code{\link{save_plots}()}
}
\concept{plot}
\concept{postprocess}
