% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cases}
\alias{plot_cases}
\title{Plot the posterior prediction for cases}
\usage{
plot_cases(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  all_obs = FALSE,
  col = NULL,
  log = TRUE
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=summarise_posterior]{summarise_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{col}{A character string denoting the variable to use to
stratify the ribbon plot. Defaults to "type" which indicates the
data stream.}

\item{log}{Logical, defaults to \code{TRUE}. Should cases be plot on
a log scale?}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the posterior prediction for cases
}
\examples{
posterior <- load_example(strains = 2, type = "posterior")

# default with log transform
plot_cases(posterior)

# without log transform
plot_cases(posterior)
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc}()}
}
\concept{plot}
