% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_delta}
\alias{plot_delta}
\title{Plot the posterior prediction for the fraction of samples with the DELTA
variant}
\usage{
plot_delta(posterior_delta, obs_delta, forecast_date = NULL)
}
\description{
Plot the posterior prediction for the fraction of samples with the DELTA
variant
}
