% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot the posterior prediction for the growth rate}
\usage{
plot_growth(posterior, forecast_dates = NULL, col = NULL)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{summarise_posterior()}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{forecast_dates}{A data.frame in the format produced by
\code{extract_forecast_dates()} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}

\item{col}{A character string denoting the variable to use to
stratify the ribbon plot. Defaults to "type" which indicates the
data stream.}
}
\description{
Plot the posterior prediction for the growth rate
}
