% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Pairs plot of parameters of interest and fitting diagnostics}
\usage{
plot_pairs(
  fit,
  pars = c("r_init", "r_noise", "beta", "delta_noise[1]", "ndelta_noise[1]",
    "init_cases", "init_cases[1]", "init_cases[2]", "eta[1]", "delta_eta[1]",
    "ndelta_eta[1]", "sqrt_phi[1]", "sqrt_phi[2]", "sqrt_phi"),
  diagnostics = TRUE,
  ...
)
}
\arguments{
\item{fit}{List of output as returned by \code{stan_fit()}.}

\item{pars}{Character vector of parameters to try and include
in the plot. Will only be included if present in the fitted model.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be shown.}

\item{...}{Additional parameters passed to \code{mcmc_pairs()}.}
}
\value{
A ggplot2 based pairs plot of parameters of interest
}
\description{
Pairs plot of parameters of interest and fitting diagnostics
}
\examples{
\dontrun{
obs <- latest_obs(germany_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
plot_pairs(fit)
}
}
