% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot the posterior prediction for the reproduction number}
\usage{
plot_rt(posterior, forecast_date = NULL)
}
\description{
Plot the posterior prediction for the reproduction number
}
