% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_voc}
\alias{plot_voc}
\title{Plot the posterior prediction for the fraction of samples with the variant
of concern}
\usage{
plot_voc(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  voc_label = "variant of concern"
)
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{obs}{A data frame of observed data as produced by \code{latest_obs()}.}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast vs an estimate. Defaults to empty in which case
forecast dates are inferred from the \code{posterior} list based on data
availability for cases and sequences. These dates can be overridden
by supplying a replacement data with a duplicate name (see the examples).}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}
}
\description{
Plot the posterior prediction for the fraction of samples with the variant
of concern
}
