% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_voc_advantage}
\alias{plot_voc_advantage}
\title{Plot the posterior prediction for the transmission advantage for the variant
of concern}
\usage{
plot_voc_advantage(
  posterior,
  forecast_dates = NULL,
  voc_label = "variant of concern",
  ...
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}

\item{...}{Additional parameters passed to \code{\link[=plot_default]{plot_default()}}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the posterior prediction for the transmission advantage for the variant
of concern
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")
plot_voc_advantage(posterior)
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc}()}
}
\concept{plot}
