% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_voc_frac}
\alias{plot_voc_frac}
\title{Plot the population posterior prediction for the fraction of samples with the
variant of concern}
\usage{
plot_voc_frac(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  all_obs = FALSE,
  central = FALSE,
  voc_label = "variant of concern",
  logit = TRUE,
  ...
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{central}{Logical, defaults to FALSE. Should the mean and median
central estimates be plot as dashed and solid lines respectively. Requires
\code{mean} and \code{median} variables to be present in the input.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}

\item{logit}{Logical, defaults to \code{TRUE}. Should variant proportions be
plot on the logit scale.}

\item{...}{Additional parameters passed to \code{\link[=plot_default]{plot_default()}}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the population posterior prediction for the fraction of samples with the
variant of concern
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")
plot_voc_frac(posterior)
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot.fv_posterior}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{save_plots}()}
}
\concept{plot}
