% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{print.fv_posterior}
\alias{print.fv_posterior}
\title{Print method for fv_tidy_posterior}
\usage{
\method{print}{fv_posterior}(x, ...)
}
\arguments{
\item{x}{An output from output from \code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}.}

\item{...}{Pass additional arguments to \code{data.table} printing method.}
}
\value{
A summary data.frame
}
\description{
\code{print} method for class "fv_posterior". Prints the available
value types and then falls back to the \code{data.table} print method.
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")

# case summary
posterior
}
\seealso{
fv_tidy_posterior

Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{plot.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
