% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{quantiles_to_long}
\alias{quantiles_to_long}
\title{Convert summarised quantiles from wide to long format}
\usage{
quantiles_to_long(posterior)
}
\arguments{
\item{posterior}{A dataframe as output by \code{summarise_posterior()},
extract_forecast(), etc.}
}
\value{
A data frame of quantiles in long format.
}
\description{
Convert summarised quantiles from wide to long format
}
\examples{
\dontrun{
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_covid19_delta_obs), max_treedepth = 15)
dt <- combine_posteriors_dt(dt, target = "forecast")
long <- quantiles_to_long(dt)
print(long)
}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{bp_launch_shinystan}()},
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{summarise_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
