% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{sample_sequences}
\alias{sample_sequences}
\title{Sample Sequence Observation Model}
\usage{
sample_sequences(frac_voc, seq_total, phi)
}
\arguments{
\item{frac_voc}{A numeric vector of expected proportions positive for the
variant of concern.}

\item{seq_total}{An integer vector of total sequences available.}

\item{phi}{The overdispersion of the sampling process. If not supplied
then no overdispersion is used (i.e a binomial observation model vs a
beta binomial observation model).}
}
\value{
A vector of observed sequences positive for the variant of
concern.
}
\description{
Sample Sequence Observation Model
}
\examples{
# dummy sequence data
frac_voc <- seq(0, 1, by = 0.1)
seq_total <- seq(10, length.out = length(frac_voc), by = 100)

# binomial observation model
sample_sequences(frac_voc, seq_total)

# beta binomial observation model
sample_sequences(frac_voc, seq_total, 0.5)
}
\seealso{
Functions to generate simulated data
\code{\link{generate_obs}()}
}
\concept{generatedata}
