% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{save_posterior}
\alias{save_posterior}
\title{Save a summarised posterior}
\usage{
save_posterior(posterior, save_path = tempdir())
}
\arguments{
\item{posterior}{A data frame of summarised posterior estimates
as returned by \code{cmdstanr::summary} with  an additional type variable
which contains the following character string options: "non-VOC",
"VOC", "Combined", "Overall".}

\item{save_path}{A character string giving the path to save the
posterior to as a series of csv's.}
}
\description{
Save a summarised posterior
}
