% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_data}
\alias{stan_data}
\title{Format data for use with stan}
\usage{
stan_data(
  obs,
  horizon = 4,
  delta = c(0.2, 0.2),
  variant_relationship = "pooled",
  likelihood = TRUE,
  output_loglikelihood = FALSE
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{delta}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}
}
\description{
Format data for use with stan
}
\examples{
stan_data(latest_obs(germany_obs))
}
