% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_fit}
\alias{stan_fit}
\title{Fit a brancing process strain model}
\usage{
stan_fit(
  data,
  model = bp.delta::load_model(strains = 2),
  save_path,
  diagnostics = TRUE,
  ...
)
}
\description{
Fit a brancing process strain model
}
\examples{
\dontrun{
# format example data
dt <- stan_data(germany_cases)

# single strain model
inits <- stan_inits(dt, strains = 1)
mod <- load_model(strains = 1)
fit <- stan_fit(dt, model = mod, init = inits, adapt_delta = 0.99)
fit

# two strain model
inits <- stan_inits(dt, strains = 2)
mod <- load_model(strains = 2)
2strain_fit <- stan_fit(dt, model = mod, init = inits, adapt_delta = 0.99)
2strain_fit
}
}
