% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_inits}
\alias{stan_inits}
\title{Set up initial conditions for model}
\usage{
stan_inits(data, strains = 2)
}
\arguments{
\item{data}{A list of data as produced by \code{stan_data()}.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}
}
\description{
Set up initial conditions for model
}
\examples{
dt <- stan_data(latest_obs(germany_covid19_delta_obs))
inits <- stan_inits(dt)
inits
inits()
}
