% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-methods.R
\name{summary.fv_forecast}
\alias{summary.fv_forecast}
\title{Summary method for forecast}
\usage{
\method{summary}{fv_forecast}(object, target = "posterior", type = "all", ...)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=forecast]{forecast()}} of class "fv_forecast".}

\item{target}{gwgfw}

\item{type}{A character string used to filter the summarised output.}

\item{...}{Pass additional arguments to \code{\link[=summary.fv_posterior]{summary.fv_posterior()}}.}
}
\value{
A summary \code{data.table}.
}
\description{
\code{summary} method for class "fv_forecast".
}
\seealso{
summary.fv_posterior forecast unnest_posterior

Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast_n_strain}()},
\code{\link{forecast}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot.fv_posterior}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
