% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{summary.fv_posterior}
\alias{summary.fv_posterior}
\title{Summary method for fv_tidy_posterior}
\usage{
\method{summary}{fv_posterior}(x, type = "cases", forecast = FALSE, as_dt = FALSE, ...)
}
\arguments{
\item{type}{A character string used to filter the summarised output.}

\item{as_dt}{Logical defaults to \code{FALSE}. Should the output}

\item{object}{An object of the class \code{fv_posterior} as returned by
\code{fv_tidy_posterior()} .}
}
\value{
A summary data.table table unless type "all" is used in which case
the output is still of type "fv_posterior"
}
\description{
\code{summary} method for class "fv_tidy_posterior". Can be used to
filter the posterior for variables of interest, to return forecats only, and
to summarise using the \code{data.table} method
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")

# case summary
summary(posterior)

# summary of the case summary
summary(posterior, as_dt = TRUE)

# case forecast only
summary(posterior, forecast = TRUE)

# voc fraction summary
summary(posterior, type = "voc_frac")

# voc advantage summary
summary(posterior, type = "voc_advantage")

# growth summary
summary(posterior, type = "growth")

# Rt summary
summary(posterior, type = "rt")

# model parameter summary
summary(posterior, type = "model")

# raw posterior values
summary(posterior, type = "raw")
}
\seealso{
fv_tidy_posterior

Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
