% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-tools.R
\name{unnest_posterior}
\alias{unnest_posterior}
\title{Unnest posterior estimates from a forecast data.frame}
\usage{
unnest_posterior(forecasts, target = "posterior")
}
\arguments{
\item{forecasts}{A data frame of forecasts as produced by \code{\link[=forecast]{forecast()}}.}

\item{target}{A character string indicating the list of outputs to
unnest.}
}
\value{
An unnested \code{data.frame} of posterior estimates and other variables
produced by \code{\link[=forecast]{forecast()}}.
}
\description{
Unnest posterior predictions and forecasts from output
produced by \code{\link[=forecast]{forecast()}} (or multiple combined calls) dropping diagnostic
and fitting variables in the process.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(data.table)
options(mc.cores = 4)
dt <- forecast(
  germany_covid19_delta_obs,
  forecast_date = as.Date("2021-06-12"),
  max_treedepth = 15, adapt_delta = 0.99
)

# unnest posterior predictions
posterior <- unnest_posterior(dt)
posterior

# unnest forecasts
forecasts <- unnest_posterior(dt, target = "forecast")
forecasts
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast_n_strain}()},
\code{\link{forecast}()},
\code{\link{plot.fv_forecast}()},
\code{\link{summary.fv_forecast}()}
}
\concept{forecast}
