% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_data_availability}
\alias{update_data_availability}
\title{Update data based on availability and forecast date}
\usage{
update_data_availability(
  cases,
  forecast_date = max(cases$date),
  cases_lag,
  seq_lag
)
}
\arguments{
\item{cases}{A data frame with the following variables:
date, cases, seq_delta, and seq_total.}

\item{forecast_date}{Date at which to forecast. Defaults to the
maximum date in \code{cases}.}

\item{cases_lag}{Number of weeks that case data takes to be reported.
Defaults to not alter the input data.}

\item{seq_lag}{Number of weeks that sequence data takes to be reported.
Defaults to not alter the input data.}
}
\description{
Update data based on availability and forecast date
}
\examples{
update_data_availability(germany_cases, cases_lag = 2, seq_lag = 3)
}
