% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenarios.R
\name{update_obs_availability}
\alias{update_obs_availability}
\title{Update observations based on availability}
\usage{
update_obs_availability(obs, cases_lag, seq_lag)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_voc, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{cases_lag}{Number of weeks that case data takes to be reported.
Defaults to not alter the input data.}

\item{seq_lag}{Number of weeks that sequence data takes to be reported.
Defaults to not alter the input data.}
}
\value{
A data frame of observations with updated case and
sequence availability dates.
}
\description{
Update observations based on availability
}
\examples{
update_obs_availability(
  germany_covid19_delta_obs,
  cases_lag = 2, seq_lag = 3
)
}
\seealso{
Functions to define and create data scenarios.
\code{\link{define_scenarios}()},
\code{\link{generate_obs_scenario}()}
}
\concept{scenarios}
