% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{update_voc_label}
\alias{update_voc_label}
\title{Label the Variant of Concern}
\usage{
update_voc_label(posterior, label, target_label = "VOC")
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=summarise_posterior]{summarise_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{label}{Character string  indicating the new label to use for the
variant of concern.}

\item{target_label}{A character string defaulting to "VOC". Indicates the
current label for the variant of concern.}
}
\value{
A list of data frames as returned by `\code{\link[=summarise_posterior]{summarise_posterior()}} but
with updated labels.
}
\description{
Assign a custom label to the variant of concern in the
output from \code{\link[=summarise_posterior]{summarise_posterior()}}.
}
\examples{
p <- fv_example(strains = 2, type = "posterior")
p <- update_voc_label(p, "Delta")
p[value_type == "model"]
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summarise_posterior}()}
}
\concept{postprocess}
