% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_forecast_dates}
\alias{add_forecast_dates}
\title{Add the forecast dates to a plot}
\usage{
add_forecast_dates(plot, forecast_dates = NULL)
}
\arguments{
\item{plot}{\code{ggplot2} object}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}
}
\value{
A \code{ggplot2} plot with dates of data unavailability added.
}
\description{
Add the forecast dates to a plot
}
\seealso{
Plotting functions
\code{\link{plot.fv_forecast}()},
\code{\link{plot.fv_posterior}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()}
}
\concept{plot}
