% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validation.R
\name{bp_launch_shinystan}
\alias{bp_launch_shinystan}
\title{Launch shinystan}
\usage{
bp_launch_shinystan(fit)
}
\arguments{
\item{fit}{List of output as returned by \code{\link[=fv_sample]{fv_sample()}}.}
}
\description{
Launch shinystan an interactive tool for stan model evaluation
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12")
)

dt <- fv_as_data_list(obs)
inits <- fv_inits(dt)
fit <- fv_sample(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
bp_launch_shinystan(fit)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions to explore and validate models
\code{\link{fv_score_forecast}()},
\code{\link{plot_pairs}()}
}
\concept{modelvalidation}
