% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validation.R
\name{bp_launch_shinystan}
\alias{bp_launch_shinystan}
\title{Launch shinystan}
\usage{
bp_launch_shinystan(fit)
}
\arguments{
\item{fit}{List of output as returned by \code{stan_fit()}.}
}
\description{
Launch shinystan an interactive tool for stan model evaluation
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
bp_launch_shinystan(fit)
}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summarise_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
