% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check a data.frame}
\usage{
check_dataframe(dataframe, req_vars, req_types, rows)
}
\arguments{
\item{dataframe}{A data.frame to check.}

\item{req_vars}{A character vector of variables that are required.}

\item{req_types}{A character vector of types for each required variable.}

\item{rows}{Integer specifying the number of rows the data.frame should have.}
}
\description{
Check a data.frame
}
\seealso{
Functions used for checking inputs.
\code{\link{check_observations}()},
\code{\link{check_param}()},
\code{\link{check_quantiles}()}
}
\concept{check}
