% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_observations}
\alias{check_observations}
\title{Check observations are in the correct format}
\usage{
check_observations(obs)
}
\arguments{
\item{obs}{A \code{data.frame} with the following variables:
\code{date}, \code{cases}, \code{seq_voc}, and \code{seq_total}, \code{cases_available},
and \code{seq_available}. \code{seq_available} and \code{case_available} must be
uniquely define data rows but other rows can be duplicated based
on data availability. This data format allows for multiple versions of
case and sequence data for a given date with different reporting dates.
This is important when using the package in evaluation settings or in
real-time where data sources are liable to be updated as new data becomes
available. See \link{germany_covid19_delta_obs} for an example of a supported
data set.}
}
\description{
Check observations are in the correct format
}
\examples{
obs <- latest_obs(germany_covid19_delta_obs)
check_observations(obs)
}
\seealso{
Functions used for checking inputs
\code{\link{check_dataframe}()},
\code{\link{check_param}()},
\code{\link{check_quantiles}()}
}
\concept{check}
