% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_param}
\alias{check_param}
\title{Check a parameter is the correct type and length}
\usage{
check_param(param, name = "param", type = "numeric", length = 1)
}
\arguments{
\item{param}{A parameter to check the format of.}

\item{name}{A character string naming the variable
to check.}

\item{type}{A character string identifying the allowed parameter
type (must be a type with an is.type function).)}

\item{length}{Numeric, allowed length of the variable.}
}
\description{
Check a parameter is the correct type and length
}
\keyword{internal}
