% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_param}
\alias{check_param}
\title{Check a parameter is the correct type and length}
\usage{
check_param(param, name = "param", type = "numeric", length)
}
\arguments{
\item{param}{A parameter to check the format of.}

\item{name}{A character string naming the variable
to check.}

\item{type}{A character string identifying the allowed parameter
type (must be a type with a is.type function except for a Date).}

\item{length}{Numeric, allowed length of the variable. Defaults to
any allowed length.}
}
\description{
Check a parameter is the correct type and length
}
\seealso{
Functions used for checking inputs.
\code{\link{check_dataframe}()},
\code{\link{check_observations}()},
\code{\link{check_quantiles}()}
}
\concept{check}
