% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{combine_posteriors}
\alias{combine_posteriors}
\title{Combine multiple summarised posteriors}
\usage{
combine_posteriors(
  posteriors_list,
  list_id = "model",
  combine_variables = FALSE,
  variable_id = "value_type"
)
}
\arguments{
\item{posteriors_list}{A list of posteriors as produced by
\code{summarise_posterior()}.}

\item{list_id}{A character string naming the variable used to identify
list parameters}

\item{combine_variables}{Logical, defaults to FALSE. Should variables
be combined into a single data frame (labelled using \code{variable_id}).}

\item{variable_id}{A character string naming the variable used to identify
variables. Defaults to \code{value_type}}
}
\description{
Combine multiple summarised posteriors
}
