% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{combine_posteriors}
\alias{combine_posteriors}
\title{Combine multiple summarised posteriors}
\usage{
combine_posteriors(posteriors_list, list_id = "model", ids)
}
\arguments{
\item{posteriors_list}{A list of posteriors as produced by
\code{summarise_posterior()}.}

\item{list_id}{A character string naming the variable used to identify
list parameters}

\item{ids}{A character vector of the same lebngth}
}
\description{
Combine multiple summarised posteriors
}
