% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{combine_posteriors_dt}
\alias{combine_posteriors_dt}
\title{Combine posteriors from a data frame}
\usage{
combine_posteriors_dt(
  forecasts,
  target = "forecast",
  combine = TRUE,
  combine_by = c("id", "forecast_date", "strains", "overdispersion",
    "variant_relationship")
)
}
\arguments{
\item{forecasts}{A data frame of forecasts as produced by \code{forecast_dt()}}

\item{target}{A character string indicating the list of posteriors to
combine. This can be either "posterior", or "forecast".}
}
\value{
A data frame of summmarised posteriors along with summary
information about the forecast method used.
}
\description{
A wrapper for \code{combine_posteriors()} but designed for use with the output
from \code{forecast_dt()}. Can be used to target either summarised posterior
estimates or forecast estimates.
}
\examples{
\dontrun{
library(data.table)
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_covid19_delta_obs), max_treedepth = 15)
dt <- rbind(dt, copy(dt)[, overdispersion := FALSE])

# combine into an unnested data.frame
dt <- combine_posteriors_dt(dt, target = "forecast")

}
}
