% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract posterior draws}
\usage{
extract_draws(fit, ...)
}
\arguments{
\item{fit}{A list as produced by \code{stan_fit()}.}

\item{...}{Additional parameters passed to \code{cmdstanr::draws}}
}
\value{
A \code{draws} object from the \code{posterior} package.
}
\description{
Extract posterior draws
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
extract_draws(fit)
}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{bp_launch_shinystan}()},
\code{\link{convert_to_stanfit}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summarise_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
