% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract posterior draws}
\usage{
extract_draws(fit, ...)
}
\arguments{
\item{fit}{A list as produced by \code{\link[=fv_sample]{fv_sample()}}.}

\item{...}{Additional parameters passed to \code{\link[cmdstanr:fit-method-draws]{cmdstanr::draws()}}}
}
\value{
A \code{\link[cmdstanr:fit-method-draws]{cmdstanr::draws()}} object from the \code{posterior} package.
}
\description{
Extract posterior draws
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- fv_as_data_list(obs)
inits <- fv_inits(dt)
fit <- fv_sample(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
extract_draws(fit)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{plot.fv_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
