% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_forecast}
\alias{extract_forecast}
\title{Extract forecasts from a summarised posterior}
\usage{
extract_forecast(posterior, forecast_dates = NULL)
}
\arguments{
\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast. Must contain a "Cases" date and a "Sequence" date.
Default is to infer these dates from the summarised posterior.}
}
\value{
A list containing a forecast for each parameter
}
\description{
Extract forecasts from a summarised posterior
}
\examples{
\dontrun{
obs <- latest_obs(germany_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
p <- summarise_posterior(fit)
extract_forecast(p)
}
}
