% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{extract_forecast_dates}
\alias{extract_forecast_dates}
\title{Extract forecast dates}
\usage{
extract_forecast_dates(posterior)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{summarise_posterior()}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}
}
\value{
A data.frame containing at least two vectors: Data unavailable
indicating the type of data that is missing, and date giving the date
data was last available for.
}
\description{
Extract forecast dates based on the availability of both case
and sequence data.
}
\examples{
\dontrun{
options(mc.cores = 4)
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs, overdispersion = FALSE)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, max_treedepth = 15, adapt_delta = 0.9)
p <- summarise_posterior(fit)

extract_forecast_dates(p)
}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{bp_launch_shinystan}()},
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summarise_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
