% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_by_availability}
\alias{filter_by_availability}
\title{Filter data based on availability and forecast date}
\usage{
filter_by_availability(
  obs,
  date = max(obs$date),
  seq_date = date,
  case_date = date
)
}
\arguments{
\item{obs}{A \code{data.frame} with the following variables:
\code{date}, \code{cases}, \code{seq_voc}, and \code{seq_total}, \code{cases_available},
and \code{seq_available}. \code{seq_available} and \code{case_available} must be
uniquely define data rows but other rows can be duplicated based
on data availability. This data format allows for multiple versions of
case and sequence data for a given date with different reporting dates.
This is important when using the package in evaluation settings or in
real-time where data sources are liable to be updated as new data becomes
available. See \link{germany_covid19_delta_obs} for an example of a supported
data set.}

\item{date}{Date at which to filter. Defaults to the
maximum date in \code{obs}.}

\item{seq_date}{Date from which to use available sequence data. Defaults to
the \code{date}.}

\item{case_date}{Date from which to use available case data. Defaults to
the \code{date}.}
}
\value{
A \code{data.frame} of observations filter for the latest available
data for the specified dates of interest.
}
\description{
Filter data based on availability and forecast date
}
\examples{
options(mc.cores = 4)
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- rbind(
  update_obs_availability(obs, seq_lag = 3),
  update_obs_availability(obs, seq_lag = 1)
)
# filter out duplicates and up to the present date
filter_by_availability(dt)

# filter to only use sequence data up the the 12th of June
filter_by_availability(dt, seq_date = "2021-06-12")

# as above but only use
filter_by_availability(dt,
  seq_date = "2021-06-12",
  case_date = "2021-07-01"
)
}
\seealso{
Preprocessing functions
\code{\link{fv_dow_period}()},
\code{\link{latest_obs}()},
\code{\link{piecewise_steps}()}
}
\concept{preprocess}
