% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast using branching processes at a target date}
\usage{
forecast(
  cases,
  target_date = max(cases$date),
  save_path = tempdir(),
  horizon = 4,
  delta = c(0.2, 0.2),
  strains = 2,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  ...
)
}
\description{
Forecast using branching processes at a target date
}
