% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast using branching processes at a target date}
\usage{
forecast(
  obs,
  forecast_date = max(obs$date),
  seq_date = forecast_date,
  case_date = forecast_date,
  save_path = NULL,
  horizon = 4,
  voc_scale = c(0, 0.2),
  strains = 2,
  variant_relationship = "pooled",
  overdispersion = TRUE,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  debug = FALSE,
  keep_fit = TRUE,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  ...
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_voc, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_date}{Date at which to forecast. Defaults to the
maximum date in \code{obs}.}

\item{seq_date}{Date from which to use available sequence data. Defaults to
the forecast date.}

\item{case_date}{Date from which to use available case data. Defaults to
the forecast date.}

\item{save_path}{Character string indicating the save path to use for results
if required. Defaults to empty meaning that nothing is saved}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{voc_scale}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant of concern.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{overdispersion}{Logical, defaults to \code{TRUE}. Should the observations
used include overdispersion.}

\item{models}{A list of models to use named by strain. If not supplied
uses the package default model for that strain.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model.extract}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the log
likelihood be output. Disabling this will speed up fitting if evaluating the
model fit is not required.}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{keep_fit}{Logical, defaults to \code{TRUE}. Should the stan model fit be
kept and returned. Dropping this can substantially reduce memory usage in
situtations where multiple models are being fit.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for.}

\item{...}{Additional parameters passed to \code{stan_fit()}.}
}
\description{
Forecast using branching processes at a target date
}
\examples{
\dontrun{
options(mc.cores = 4)
results <- forecast(
  latest_obs(germany_covid19_delta_obs),
  horizon = 4,
  save_path = tempdir(),
  strains = c(1, 2),
  adapt_voc = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)
# inspect object
names(results)
}
}
