% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast using branching processes at a target date}
\usage{
forecast(
  obs,
  forecast_date = max(obs$date),
  seq_date = forecast_date,
  case_date = forecast_date,
  horizon = 4,
  r_init = c(0, 0.25),
  voc_scale = c(0, 0.2),
  voc_label = "VOC",
  strains = 2,
  variant_relationship = "pooled",
  overdispersion = TRUE,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  debug = FALSE,
  keep_fit = TRUE,
  scale_r = 1,
  probs = c(0.05, 0.2, 0.8, 0.95),
  id = 0,
  ...
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_voc, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_date}{Date at which to forecast. Defaults to the
maximum date in \code{obs}.}

\item{seq_date}{Date from which to use available sequence data. Defaults to
the forecast date.}

\item{case_date}{Date from which to use available case data. Defaults to
the forecast date.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{r_init}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growht rate.}

\item{voc_scale}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant of concern.}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{strains}{Integer number of strains to use. Defaults to 2. Current
maximum is 2. A numeric vector can be passed if forecasts from multiple
strain models are desired.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{overdispersion}{Logical, defaults to \code{TRUE}. Should the observations
used include overdispersion.}

\item{models}{A model as supplied by \code{load_model()}. If not supplied the
default for that strain is used. If multiple strain models are being forecast
then \code{models} should be a list models.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model.extract.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the log
likelihood be output. Disabling this will speed up fitting if evaluating the
model fit is not required.}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{keep_fit}{Logical, defaults to \code{TRUE}. Should the stan model fit be
kept and returned. Dropping this can substantially reduce memory usage in
situtations where multiple models are being fit.}

\item{scale_r}{Numeric, defaults to 1. Rescale the timespan over which
the growth rate and reproduction number is calculated. An example use case
is rescaling the growth rate from weekly to be scaled by the mean of
the generation time (for COVID-19 for example this would be 5.5 / 7.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{plot_cases()}, \code{plot_rt},
and \code{plot_voc}).}

\item{id}{ID to assign to this forecast. Defaults to 0.}

\item{...}{Additional parameters passed to \code{stan_fit()}.}
}
\value{
A dataframe containing the output of \code{stan_fit()} in each row as
well as the summarised posterior, forecast and information about the
parameters specified.
}
\description{
Forecast using branching processes at a target date
}
\examples{
\dontrun{
options(mc.cores = 4)
results <- forecast(
  latest_obs(germany_covid19_delta_obs),
  horizon = 4,
  strains = c(1, 2),
  adapt_delta = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)
# inspect results
results

# unnest posteriors
results <- unnest_posterior(results)

# plot case posterior predictions
plot_cases(results, log = TRUE)
}
}
\seealso{
Functions used for forecastin across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
