% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.vocs-package.R
\docType{package}
\name{forecast.vocs-package}
\alias{forecast.vocs}
\alias{forecast.vocs-package}
\title{forecast.vocs: Forecast Case and Sequence Notifications using Variant of Concern Strain Dynamics}
\description{
Contains models and tools to produce short-term forecasts for both case and sequence notifications assuming circulation of either one or two variants. Tools are also provided to allow the evaluation of the use of sequence data for short-term forecasts in both real-world settings and in user generated scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/forecast.vocs/}
  \item \url{https://github.com/epiforecasts/forecast.vocs/}
  \item Report bugs at \url{https://github.com/epiforecasts/forecast.vocs/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Other contributors:
\itemize{
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [contributor]
}

}
\keyword{internal}
