% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-tools.R
\name{forecast_across_dates}
\alias{forecast_across_dates}
\title{Forecast across multiple dates}
\usage{
forecast_across_dates(
  obs,
  forecast_dates = unique(obs[!is.na(seq_available)])$date[-c(1:3)],
  ...
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_voc, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_dates}{A list of dates to forecast at.}

\item{...}{Additional parameters passed to \code{forecast()}}
}
\value{
A data table each row containing the output from running
\code{forecast()} on a single forecast date.
}
\description{
Forecast across multiple dates
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
options(mc.cores = 4)

forecasts <- forecast_across_dates(
  germany_covid19_delta_obs,
  forecast_dates = c(as.Date("2021-05-01"), as.Date("2021-06-12")),
  horizon = 4,
  strains = 2,
  adapt_delta = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)

# inspect forecasts
forecasts

# unnest posteriors
posteriors <- unnest_posterior(forecasts)

# plot case posterior predictions
plot_cases(posteriors, log = TRUE) +
  facet_grid(vars(forecast_date), vars(voc_scale))
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for forecastin across models, dates, and scenarios
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
