% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{forecast_dt}
\alias{forecast_dt}
\title{Forecast using branching processes at a target date returning a
summary data frame}
\usage{
forecast_dt(
  obs,
  forecast_date = max(obs$date),
  strains = 1,
  overdispersion = TRUE,
  variant_relationship = "pooled",
  model = bp.delta::load_model(strains = strains),
  keep_forecast = FALSE,
  id = 0,
  ...
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total, cases_available,
and seq_available. seq_available and case_available must be
uniquely define data rows but other rows can be duplicated based
on data availability.}

\item{forecast_date}{Date at which to forecast. Defaults to the
maximum date in \code{obs}.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{model}{A model as loaded by \code{load_model()}.}

\item{keep_forecast}{Logical, defaults to \code{FALSE}. Should the \code{forecast()}
output be kept.}

\item{id}{An integer, defaulting to 0 to identify this forecast.}

\item{...}{Additional arguments passed to \code{forecast()}}
}
\value{
A dataframe containing the output of \code{forecast()} in each row as
well as summary information about the forecast
}
\description{
Designed to streamline workflows with a large number of
indepedent model fits this function returns a data frame summarising
a forecast for a single model (rather than the multiple models that may
be used directly with \code{forecast_dt()}). Unlike \code{forecast()} it is also
error tolerant and returns the error if forecasting fails.
}
\examples{
\dontrun{
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_obs), max_treedepth = 15)
print(dt)
}
}
