% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_n_strain}
\alias{forecast_n_strain}
\title{Forecast using a single branching process}
\usage{
forecast_n_strain(
  data,
  model = NULL,
  strains = 2,
  save_path = tempdir(),
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  ...
)
}
\description{
Forecast using a single branching process
}
