% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_n_strain}
\alias{forecast_n_strain}
\title{Forecast for a single model and summarise}
\usage{
forecast_n_strain(
  data,
  model = NULL,
  strains = 2,
  voc_label = "VOC",
  probs = c(0.05, 0.2, 0.8, 0.95),
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{stan_data()}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{load_model()}}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{plot_cases()}, \code{plot_rt},
and \code{plot_voc}).}

\item{...}{Additional parameters passed to \code{stan_fit()}.}
}
\description{
Forecast for a single model and summarise
}
