% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_n_strain}
\alias{forecast_n_strain}
\title{Forecast for a single model and summarise}
\usage{
forecast_n_strain(
  data,
  model = NULL,
  inits = forecast.vocs::fv_inits,
  fit = forecast.vocs::fv_sample,
  posterior = forecast.vocs::fv_tidy_posterior,
  extract_forecast = forecast.vocs::fv_extract_forecast,
  strains = 2,
  voc_label = "VOC",
  probs = c(0.05, 0.2, 0.8, 0.95),
  digits = 3,
  scale_r = 1,
  timespan = 7,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=fv_as_data_list]{fv_as_data_list()}}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=fv_model]{fv_model()}}.}

\item{inits}{A function that returns a function to samples initial
conditions with the same arguments as \code{\link[=fv_inits]{fv_inits()}}. If not supplied the
package default \code{\link[=fv_inits]{fv_inits()}} is used.}

\item{fit}{A function that fits the supplied model with the same arguments
and return values as \code{\link[=fv_sample]{fv_sample()}}. If not supplied the
package default \code{\link[=fv_sample]{fv_sample()}} is used which performs MCMC sampling using
\link{cmdstanr}.}

\item{posterior}{A function that summarises the output from the supplied
fitting function with the same arguments and return values (depending on
the requirement for downstream package functionality to function) as
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. If not supplied the package default
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}} is used.}

\item{extract_forecast}{A function that extracts the forecast from
the summarised \code{posterior}. If not supplied the package default
\code{\link[=fv_extract_forecast]{fv_extract_forecast()}} is used.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{\link[=plot_cases]{plot_cases()}}, \code{\link[=plot_rt]{plot_rt()}},
and \code{\link[=plot_voc_frac]{plot_voc_frac()}}).}

\item{digits}{Numeric, defaults to 3. Number of digits to round summary
statistics to.}

\item{scale_r}{Numeric, defaults to 1. Rescale the timespan over which
the growth rate and reproduction number is calculated. An example use case
is rescaling the growth rate from weekly to be scaled by the mean of
the generation time (for COVID-19 for example this would be 5.5 / 7.}

\item{timespan}{Integer, defaults to 7. Indicates the number of days between
each observation. Defaults to a week.}

\item{...}{Additional parameters passed to \code{\link[=fv_sample]{fv_sample()}}.}
}
\description{
Forecast for a single model and summarise
}
\seealso{
Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast}()},
\code{\link{plot.fv_forecast}()},
\code{\link{summary.fv_forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
