% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_n_strain}
\alias{forecast_n_strain}
\title{Forecast for a single model and summarise}
\usage{
forecast_n_strain(
  data,
  model = NULL,
  strains = 2,
  voc_label = "VOC",
  probs = c(0.05, 0.2, 0.8, 0.95),
  scale_r = 1,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=stan_data]{stan_data()}}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=load_model]{load_model()}}.}

\item{strains}{Integer number of strains. Defaults to 2. Current
maximum is 2.}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{\link[=plot_cases]{plot_cases()}}, \code{\link[=plot_rt]{plot_rt()}},
and \code{\link[=plot_voc]{plot_voc()}}).}

\item{scale_r}{Numeric, defaults to 1. Rescale the timespan over which
the growth rate and reproduction number is calculated. An example use case
is rescaling the growth rate from weekly to be scaled by the mean of
the generation time (for COVID-19 for example this would be 5.5 / 7.}

\item{...}{Additional parameters passed to \code{\link[=stan_fit]{stan_fit()}}.}
}
\description{
Forecast for a single model and summarise
}
\seealso{
Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
