% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fv_as_data_list}
\alias{fv_as_data_list}
\title{Format data for use with stan}
\usage{
fv_as_data_list(
  obs,
  horizon = 4,
  r_init = c(0, 0.25),
  voc_scale = c(0, 0.2),
  variant_relationship = "pooled",
  overdispersion = TRUE,
  likelihood = TRUE,
  output_loglik = TRUE,
  debug = FALSE
)
}
\arguments{
\item{obs}{A data frame with the following variables:
\code{date}, \code{cases}, \code{seq_voc}, and \code{seq_total}.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{r_init}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate.}

\item{voc_scale}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant of concern.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{overdispersion}{Logical, defaults to \code{TRUE}. Should the observations
used include overdispersion.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}
}
\value{
A list as required by stan.
}
\description{
Format data for use with stan
}
\examples{
fv_as_data_list(latest_obs(germany_covid19_delta_obs))
}
\seealso{
Functions used for modelling
\code{\link{fv_inits}()},
\code{\link{fv_model}()},
\code{\link{fv_sample}()}
}
\concept{model}
