% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_obs}
\alias{generate_obs}
\title{Generate Simulated Observations}
\usage{
generate_obs(
  obs,
  strains = 2,
  model = forecast.vocs::fv_model(strains = strains),
  type = "prior",
  datasets = 10,
  ...
)
}
\arguments{
\item{obs}{Observed data to use to parameterise the model and used for
fitting when the posterior is required.}

\item{strains}{Integer number of strains to use. Defaults to 2. Current
maximum is 2. A numeric vector can be passed if forecasts from multiple
strain models are desired.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=fv_model]{fv_model()}}.}

\item{type}{A character string indicating the type of data to generate.
Supported options are data based on the "prior" or data based on  the
"posterior" with the default being the prior.}

\item{datasets}{Numeric, defaults to 10. Number of datasets to generate.}

\item{...}{Additional arguments to pass \code{fv_data()}.}
}
\value{
A dataframe with a sampled dataset on each row with the following
variables: parameters (prior/posterior parameters used to generate the data),
obs (simulated observed data), fv_data, (the simulated data formatted
using \code{fv_data()} using the same arguments as specified  for simulation.)
}
\description{
Generate simulated observations from the prior or posterior
distributions of a \code{forecast.vocs} model.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)
obs <- latest_obs(germany_covid19_delta_obs)

sim_obs <- generate_obs(obs, voc_scale = c(0.8, 0.1), r_init = c(-0.1, 0.05))

# fit a simulated dataset
sim_dt <- sim_obs$fv_data[[1]]
inits <- fv_inits(sim_dt)
fit <- fv_sample(
  sim_dt,
  init = inits, adapt_delta = 0.95, max_treedepth = 15
)

# summarise and plot simualated fit
posterior <- fv_tidy_posterior(fit)

plot_cases(posterior, log = TRUE)

plot_voc(posterior)

plot_rt(posterior)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions to generate simulated data
\code{\link{sample_sequences}()}
}
\concept{generatedata}
