% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{latest_obs}
\alias{latest_obs}
\title{Filter for latest observations of all types}
\usage{
latest_obs(obs)
}
\arguments{
\item{obs}{A \code{data.frame} with the following variables:
\code{date}, \code{cases}, \code{seq_voc}, and \code{seq_total}, \code{cases_available},
and \code{seq_available}. \code{seq_available} and \code{case_available} must be
uniquely define data rows but other rows can be duplicated based
on data availability. This data format allows for multiple versions of
case and sequence data for a given date with different reporting dates.
This is important when using the package in evaluation settings or in
real-time where data sources are liable to be updated as new data becomes
available. See \link{germany_covid19_delta_obs} for an example of a supported
data set.}
}
\value{
A \code{data.frame} of observations filtered for the
latest available data.
}
\description{
Filter for latest observations of all types
}
\examples{
dt <- rbind(
  update_obs_availability(germany_covid19_delta_obs, seq_lag = 3),
  update_obs_availability(germany_covid19_delta_obs, seq_lag = 1)
)
latest_obs(dt)
}
\seealso{
Preprocessing functions
\code{\link{filter_by_availability}()}
}
\concept{preprocess}
