% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{link_dates_with_posterior}
\alias{link_dates_with_posterior}
\title{Link dates by time for posterior parameter estimates}
\usage{
link_dates_with_posterior(posterior, data, mod_end = 0)
}
\arguments{
\item{posterior}{A data frame of summarised posterior estimates
as returned by \code{cmdstanr::summary} with  an additional type variable
which contains the following character string options: "non-VOC",
"VOC", "Combined", "Overall".}

\item{data}{A list of data as returned in the "data" entry of the output
returned by \code{stan_fit()}.}

\item{mod_end}{Integer, defaults to 0. Amount to shift the end date of
estimates.}
}
\value{
A dataframe with an additional data column
}
\description{
Link dates by time for posterior parameter estimates
}
