% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_model}
\alias{load_model}
\title{Load and compile a strain model}
\usage{
load_model(strains = 2)
}
\description{
Load and compile a strain model
}
\examples{
\dontrun{
# one strain model
mod <- load_model(strains = 1)

# two strain model
two_strain_mod <- load_model(strains = 2)
}
}
