% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{piecewise_steps}
\alias{piecewise_steps}
\title{Calculate piecewise steps}
\usage{
piecewise_steps(t, step)
}
\arguments{
\item{t}{Integer, the timespan over which to calculate steps}

\item{step}{Integer, the frequency at which to step.}
}
\value{
A list containing two elements: \code{n} (the number of steps) and \code{steps}
the location of steps as a binary variable.
}
\description{
This helper function streamlines the calculation of piecewise steps.
This may be useful when specifying random walks, AR processes, etc.
}
\seealso{
Preprocessing functions
\code{\link{filter_by_availability}()},
\code{\link{latest_obs}()}
}
\concept{preprocess}
