% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{plot.fv_posterior}
\alias{plot.fv_posterior}
\title{Plot method for fv_tidy_posterior}
\usage{
\method{plot}{fv_posterior}(
  x,
  obs = NULL,
  type = "cases",
  forecast_dates = NULL,
  central = FALSE,
  all_obs = FALSE,
  voc_label = "variant of concern",
  ...
)
}
\arguments{
\item{x}{A \code{data.table} of output as produced by \code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}.}

\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{type}{A character string indicating the type of plot required,
defaulting to "cases". Current options are: "cases" which calls
\code{\link[=plot_cases]{plot_cases()}}, "voc_frac" which calls \code{\link[=plot_voc_frac]{plot_voc_frac()}}, "voc_advantage"
which calls \code{\link[=plot_voc_advantage]{plot_voc_advantage()}}, "growth" which calls \code{\link[=plot_growth]{plot_growth()}},
"rt" which calls \code{\link[=plot_rt]{plot_rt()}}, and "all" which produces a list of all plots
by call \code{\link[=plot_posterior]{plot_posterior()}}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}

\item{central}{Logical, defaults to FALSE. Should the mean and median
central estimates be plot as dashed and solid lines respectively. Requires
\code{mean} and \code{median} variables to be present in the input.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}

\item{...}{Pass additional arguments to lower level plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "fv_posterior". This function wraps all
lower level plot functions.
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")

# plot cases on the log scale
plot(posterior, type = "cases", log = TRUE)

# plot cases with central estimates
plot(posterior, type = "cases", log = FALSE, central = TRUE)

# plot fraction that have the variant of concern
plot(posterior, type = "voc_frac")

# plot the transmission advantage for the the variant of concern
plot(posterior, type = "voc_advantage")

# plot the growth rates for both voc and non-voc cases
plot(posterior, type = "growth")

# plot the reproduction number estimates
plot(posterior, type = "rt")
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{fv_tidy_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{print.fv_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{summary.fv_posterior}()},
\code{\link{update_voc_label}()}

Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()},
\code{\link{save_plots}()}
}
\concept{plot}
\concept{postprocess}
