% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cases}
\alias{plot_cases}
\title{Plot the posterior prediction for cases}
\usage{
plot_cases(posterior, obs = NULL, forecast_dates = NULL, log = TRUE)
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{obs}{A data frame of observed data as produced by \code{latest_obs()}.}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast vs an estimate. Defaults to empty in which case
forecast dates are inferred from the \code{posterior} list based on data
availability for cases and sequences. These dates can be overridden
by supplying a replacement data with a duplicate name (see the examples).}
}
\description{
Plot the posterior prediction for cases
}
