% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_default}
\alias{plot_default}
\title{Default posterior plot}
\usage{
plot_default(
  posterior,
  target,
  obs = NULL,
  forecast_dates = NULL,
  central = FALSE,
  all_obs = FALSE,
  ...
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{target}{A character string indicating which variable to extract
from the posterior list.}

\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}

\item{central}{Logical, defaults to FALSE. Should the mean and median
central estimates be plot as dashed and solid lines respectively. Requires
\code{mean} and \code{median} variables to be present in the input.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}}}
}
\value{
A \code{ggplot2} plot.
}
\description{
Default posterior plot
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot.fv_posterior}()},
\code{\link{plot_cases}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()},
\code{\link{save_plots}()}
}
\concept{plot}
