% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot the posterior prediction for the growth rate}
\usage{
plot_growth(posterior, forecast_dates = NULL, col = NULL)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=fv_tidy_posterior]{fv_tidy_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain faceting variables.}

\item{col}{A character string denoting the variable to use to
stratify the ribbon plot. Defaults to "type" which indicates the
data stream.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the posterior prediction for the growth rate
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")
plot_growth(posterior)
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc}()}
}
\concept{plot}
