% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior predictions}
\usage{
plot_posterior(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  all_obs = FALSE,
  save_path = NULL,
  type = "png"
)
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{obs}{A data frame of observed data as produced by \code{latest_obs()}.}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast vs an estimate. Defaults to empty in which case
forecast dates are inferred from the \code{posterior} list based on data
availability for cases and sequences. These dates can be overridden
by supplying a replacement data with a duplicate name (see the examples).}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}
}
\description{
Plot posterior predictions
}
\examples{
\dontrun{
obs <- latest_obs(germany_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
posterior <- summarise_posterior(fit)
plot_posterior(posterior)
}
}
