% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior predictions}
\usage{
plot_posterior(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  save_path = NULL,
  type = "png",
  voc_label = "variant of concern"
)
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{obs}{A data frame of observed data as produced by \code{latest_obs()}.}

\item{forecast_dates}{A named vector of dates to use to identify when
output is a forecast vs an estimate. Defaults to empty in which case
forecast dates are inferred from the \code{posterior} list based on data
availability for cases and sequences. These dates can be overridden
by supplying a replacement data with a duplicate name (see the examples).}

\item{save_path}{A character string indicating where to save plots
if required.}

\item{type}{A character string indicating the format to use to save plots.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}
}
\description{
Plot posterior predictions
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
posterior <- summarise_posterior(fit)
plot_posterior(posterior)
}
}
