% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior predictions}
\usage{
plot_posterior(posterior, cases, forecast_date = NULL, save_path, type = "png")
}
\description{
Plot posterior predictions
}
\examples{
\dontrun{
dt <- stan_data(germany_cases)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
posterior <- summarise_posterior(fit)
plot_posterior(posterior, germany_cases)
}
}
