% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior predictions}
\usage{
plot_posterior(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  save_path = NULL,
  type = "png",
  all_obs = FALSE,
  voc_label = "variant of concern"
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{\link[=summarise_posterior]{summarise_posterior()}}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{\link[=extract_forecast_dates]{extract_forecast_dates()}} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}

\item{save_path}{A character string indicating where to save plots
if required.}

\item{type}{A character string indicating the format to use to save plots.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}
}
\value{
A named list of all supported package plots with sensible defaults.
}
\description{
Plot posterior predictions
}
\examples{
posterior <- load_example(strains = 2, type = "posterior")
plot_posterior(posterior)
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc}()}
}
\concept{plot}
