% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior predictions}
\usage{
plot_posterior(posterior, obs, forecast_dates = NULL, save_path, type = "png")
}
\description{
Plot posterior predictions
}
\examples{
\dontrun{
obs <- latest_obs(germany_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
posterior <- summarise_posterior(fit)
plot_posterior(posterior, obs)
}
}
