% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot the posterior prediction for the reproduction number}
\usage{
plot_rt(posterior, forecast_dates = NULL)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{summarise_posterior()}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{forecast_dates}{A data.frame in the format produced by
\code{extract_forecast_dates()} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}
}
\description{
Plot the posterior prediction for the reproduction number
}
