% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot the posterior prediction for the reproduction number}
\usage{
plot_rt(posterior, forecast_dates = NULL, col = NULL)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{summarise_posterior()}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{forecast_dates}{A data.frame in the format produced by
\code{extract_forecast_dates()} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}

\item{col}{A character string denoting the variable to use to
stratify the ribbon plot. Defaults to "type" which indicates the
data stream.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the posterior prediction for the reproduction number
}
\seealso{
Plotting functionts
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc}()}
}
\concept{plot}
