% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_voc}
\alias{plot_voc}
\title{Plot the posterior prediction for the fraction of samples with the variant
of concern}
\usage{
plot_voc(
  posterior,
  obs = NULL,
  forecast_dates = NULL,
  all_obs = FALSE,
  voc_label = "variant of concern",
  ...
)
}
\arguments{
\item{posterior}{A dataframe of posterior output as produced by
\code{summarise_posterior()}. For forecast dates to be extracted data with
\code{value_type == "cases"} must be present.}

\item{obs}{A data frame of observed data as produced by \code{latest_obs()}.}

\item{forecast_dates}{A data.frame in the format produced by
\code{extract_forecast_dates()} (with at least a date variable and a
Data unavailable variable)). Specifies when date availability should be
add to plots. May contain facetting variables.}

\item{all_obs}{Logical, defaults to \code{FALSE}. Should all observations be plot
or just those in the date range of the estimates being plot.}

\item{voc_label}{Character string giving the name to assign to the variant
of concern. Defaults to  "variant of concern".}

\item{...}{Additional parameters passed to \code{plot_default()}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot the posterior prediction for the fraction of samples with the variant
of concern
}
\seealso{
Plotting functionts
\code{\link{add_forecast_dates}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()}
}
\concept{plot}
