% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{quantiles_to_long}
\alias{quantiles_to_long}
\title{Convert summarised quantiles from wide to long format}
\usage{
quantiles_to_long(posterior)
}
\arguments{
\item{posterior}{A dataframe as output by \code{\link[=fv_posterior]{fv_posterior()}},
\code{\link[=fv_extract_forecast]{fv_extract_forecast()}}, etc.}
}
\value{
A data frame of quantiles in long format.
}
\description{
Convert summarised quantiles from wide to long format
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")
long_posterior <- quantiles_to_long(posterior)
long_posterior
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{fv_extract_forecast}()},
\code{\link{fv_posterior}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
