% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{quantiles_to_long}
\alias{quantiles_to_long}
\title{Convert summarised quantiles from wide to long format}
\usage{
quantiles_to_long(posterior)
}
\arguments{
\item{posterior}{A dataframe as output by \code{\link[=summarise_posterior]{summarise_posterior()}},
\code{\link[=extract_forecast]{extract_forecast()}}, etc.}
}
\value{
A data frame of quantiles in long format.
}
\description{
Convert summarised quantiles from wide to long format
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- forecast(obs, max_treedepth = 15, adapt_delta = 0.99)
posterior <- unnest_posterior(dt)
long_posterior <- quantiles_to_long(posterior)
long_posterior
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{bp_launch_shinystan}()},
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{summarise_posterior}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
