% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantiles_to_long}
\alias{quantiles_to_long}
\title{Convert summarised quantiles from wide to long format}
\usage{
quantiles_to_long(posterior)
}
\arguments{
\item{posterior}{A dataframe as output by \code{summarise_posterior()},
extract_forecast(), combine_posteriors_dt(), etc.}
}
\value{
A data frame of quantiles in long format.
}
\description{
Convert summarised quantiles from wide to long format
}
\examples{
\dontrun{
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_obs), max_treedepth = 15)
dt <- combine_posteriors_dt(dt, target = "forecast")
long <- quantiles_to_long(dt)
print(long)
}
}
