% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plots}
\alias{save_plots}
\title{Save plots by name}
\usage{
save_plots(plots, save_path = NULL, type = "png", ...)
}
\arguments{
\item{plots}{A named list of \code{ggplot2} plots.}
}
\description{
Save plots by name
}
\examples{
posterior <- fv_example(strains = 2, type = "posterior")
p <- plot(posterior, type = "all")
save_plots(p, save_path = tempdir())
}
\seealso{
Plotting functions
\code{\link{add_forecast_dates}()},
\code{\link{plot.fv_forecast}()},
\code{\link{plot.fv_posterior}()},
\code{\link{plot_cases}()},
\code{\link{plot_default}()},
\code{\link{plot_growth}()},
\code{\link{plot_pairs}()},
\code{\link{plot_posterior}()},
\code{\link{plot_rt}()},
\code{\link{plot_theme}()},
\code{\link{plot_voc_advantage}()},
\code{\link{plot_voc_frac}()}
}
\concept{plot}
