% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_data}
\alias{stan_data}
\title{Format data for use with stan}
\usage{
stan_data(
  cases,
  horizon = 4,
  delta = c(0.2, 0.2),
  likelihood = TRUE,
  output_loglikelihood = FALSE
)
}
\arguments{
\item{cases}{A data frame with the following variables:
date, cases, seq_delta, and seq_total.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{delta}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant.}
}
\description{
Format data for use with stan
}
\examples{
stan_data(germany_cases)
}
