% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_data}
\alias{stan_data}
\title{Format data for use with stan}
\usage{
stan_data(
  cases,
  horizon = 4,
  delta = c(0.2, 0.2),
  likelihood = TRUE,
  output_loglikelihood = FALSE
)
}
\description{
Format data for use with stan
}
\examples{
stan_data(germany_cases)
}
