% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_data}
\alias{stan_data}
\title{Format data for use with stan}
\usage{
stan_data(
  obs,
  horizon = 4,
  delta = c(0.2, 0.2),
  variant_relationship = "pooled",
  overdispersion = TRUE,
  likelihood = TRUE,
  output_loglikelihood = TRUE
)
}
\arguments{
\item{obs}{A data frame with the following variables:
date, cases, seq_delta, and seq_total.}

\item{horizon}{Integer forecast horizon. Defaults to 4.}

\item{delta}{Numeric vector of length 2. Prior mean and
standard deviation for the initial growth rate modifier
due to the variant.}

\item{variant_relationship}{Character string, defaulting to "pooled".
Controls the relationship of strains with options being "pooled" (dependence
determined from the data), "scaled" (a fixed scaling between strains), and
"independent" (fully independent strains after initial scaling).}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model.extract}

\item{output_loglikelihood}{Logical, defaults to \code{FALSE}. Should the log
likelihood be output. Disabling this will speed up fitting if evaluating the
model fit is not required.}

\item{overdisperion}{Logical, defaults to \code{TRUE}. Should the observations
used include overdispersion.}
}
\description{
Format data for use with stan
}
\examples{
stan_data(latest_obs(germany_covid19_delta_obs))
}
