% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{stan_fit}
\alias{stan_fit}
\title{Fit a brancing process strain model}
\usage{
stan_fit(
  data,
  model = forecast.vocs::load_model(strains = 2),
  diagnostics = TRUE,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by \code{stan_data()}}

\item{model}{A \code{cmdstanr} model object as loaded by \code{load_model()}}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a data frame.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\description{
Fit a brancing process strain model
}
\examples{
\dontrun{
# parallisation
options(mc.cores = 4)
# format example data
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)

# single strain model
inits <- stan_inits(dt, strains = 1)
mod <- load_model(strains = 1)
fit <- stan_fit(dt,
  model = mod, init = inits, adapt_delta = 0.99,
  max_treedepth = 15
)
fit

# two strain model
inits <- stan_inits(dt, strains = 2)
mod <- load_model(strains = 2)
two_strain_fit <- stan_fit(dt,
  model = mod, init = inits,
  adapt_delta = 0.99, max_treedepth = 15
)
two_strain_fit
}
}
