% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{summarise_posterior}
\alias{summarise_posterior}
\title{Summarise the posterior}
\usage{
summarise_posterior(
  fit,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)
)
}
\arguments{
\item{fit}{List of output as returned by \code{stan_fit()}.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for.}
}
\description{
Summarise the posterior
}
\examples{
\dontrun{
dt <- stan_data(latest_obs(germany_covid19_delta_obs))
inits <- stan_inits(dt)
options(mc.cores = 4)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
summarise_posterior(fit)
}
}
