% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{summarise_posterior}
\alias{summarise_posterior}
\title{Summarise the posterior}
\usage{
summarise_posterior(
  fit,
  probs = c(0.05, 0.2, 0.8, 0.95),
  voc_label = "VOC",
  scale_r = 1
)
}
\arguments{
\item{fit}{List of output as returned by \code{\link[=stan_fit]{stan_fit()}}.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work (such as \code{\link[=plot_cases]{plot_cases()}}, \code{\link[=plot_rt]{plot_rt()}},
and \code{\link[=plot_voc]{plot_voc()}}).}

\item{voc_label}{A character string, default to "VOC". Defines the label
to assign to variant of concern specific parameters. Example usage is to
rename parameters to use variant specific terminology.}

\item{scale_r}{Numeric, defaults to 1. Rescale the timespan over which
the growth rate and reproduction number is calculated. An example use case
is rescaling the growth rate from weekly to be scaled by the mean of
the generation time (for COVID-19 for example this would be 5.5 / 7.}
}
\value{
A dataframe summarising the model posterior. Output is stratified
by \code{value_type} with posterior summaries by case, voc, rt, growth, model,
and the raw posterior summary.
}
\description{
Summarise the posterior
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
summarise_posterior(fit)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{bp_launch_shinystan}()},
\code{\link{convert_to_stanfit}()},
\code{\link{extract_draws}()},
\code{\link{extract_forecast_dates}()},
\code{\link{extract_forecast}()},
\code{\link{link_dates_with_posterior}()},
\code{\link{link_obs_with_posterior}()},
\code{\link{quantiles_to_long}()},
\code{\link{update_voc_label}()}
}
\concept{postprocess}
