% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-methods.R
\name{summary.fv_forecast}
\alias{summary.fv_forecast}
\title{Summary method for forecast}
\usage{
\method{summary}{fv_forecast}(
  object,
  target = "posterior",
  type = "model",
  as_dt = FALSE,
  forecast = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=forecast]{forecast()}} of class "fv_forecast".}

\item{target}{A character string indicating the target object within the
\code{\link[=forecast]{forecast()}} to summarise. Current options are: posterior predictions
("posterior"), posterior forecasts ("forecast"), the model fit ("fit"),
and the model diagnostics ("diagnostics"). When "posterior" or "forecast"
are used then \code{\link[=summary.fv_posterior]{summary.fv_posterior()}} is called on the nested posterior or
forecast.}

\item{type}{A character string used to filter the summarised output and
defaulting to "model". Current options are: "model" which returns a
summary of key model parameters, "cases" which returns summarised cases,
"voc_frac" which returns summarised estimates of the fraction of cases that
have the variant of concern, "voc_advantage" that returns summarised
estimates of the the transmission advantage of the variant of concern,
"growth" which returns summarised variant specific and overall growth rates,
"rt" which returns summarised variant specific and overall reproduction
number estimates, "raw" which returns a raw posterior summary, and "all"
which returns all tidied posterior estimates.}

\item{as_dt}{Logical defaults to \code{FALSE}. Once any filtering has been applied
should \code{\link[=summary]{summary()}} fall back to using the default \code{data.table} method.}

\item{forecast}{Logical defaults to \code{FALSE}. Should \code{\link[=fv_extract_forecast]{fv_extract_forecast()}}
be used to return only forecasts rather than complete posterior.}

\item{...}{Additional summary arguments.}
}
\value{
A summary \code{data.table}.
}
\description{
\code{summary} method for class "fv_forecast".
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options(mc.cores = 4)

forecasts <- forecast(
  germany_covid19_delta_obs,
  forecast_date = as.Date("2021-06-12"),
  horizon = 4,
  strains = c(1, 2),
  adapt_delta = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)
# inspect forecasts
forecasts

# extract the model summary
summary(forecasts, type = "model")

# extract the fit object
summary(forecasts, target = "fit")

# extract the case forecast
summary(forecasts, type = "cases", forecast = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
summary.fv_posterior forecast unnest_posterior

Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast_n_strain}()},
\code{\link{forecast}()},
\code{\link{plot.fv_forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
