% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-methods.R
\name{summary.fv_forecast}
\alias{summary.fv_forecast}
\title{Summary method for forecast}
\usage{
\method{summary}{fv_forecast}(object, type = "posterior", ...)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=forecast]{forecast()}}.}

\item{type}{A character string indicating the type of summary to return.
Currently supported options are "nowcast" which summaries the nowcast
posterior using \code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}}, "nowcast_samples" which returns
posterior samples from the most recent nowcast, "fit" which returns the
summarised \code{cmdstanr} fit using \code{\link[=enw_posterior]{enw_posterior()}}, and
"posterior_prediction" which returns summarised posterior predictions for
observations used in fitting (using \code{\link[=enw_pp_summary]{enw_pp_summary()}}).}

\item{...}{Pass additional arguments to summary functions.}
}
\value{
A summary data.frame
}
\description{
\code{summary} method for class "fv_forecast".
}
\seealso{
summary forecast unnest_posterior

Functions used for forecasting across models, dates, and scenarios
\code{\link{forecast_across_dates}()},
\code{\link{forecast_across_scenarios}()},
\code{\link{forecast_n_strain}()},
\code{\link{forecast}()},
\code{\link{plot.fv_forecast}()},
\code{\link{unnest_posterior}()}
}
\concept{forecast}
