% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-tools.R
\name{unnest_posterior}
\alias{unnest_posterior}
\title{Unnest posterior estimates from a forecast data.frame}
\usage{
unnest_posterior(forecasts, target = "posterior")
}
\arguments{
\item{forecasts}{A data frame of forecasts as produced by \code{forecast()}.}

\item{target}{A character string indicating the list of outpuuts to
unnest. This can currently be either "posterior", or "forecast".}
}
\value{
An unnested data.frame of posterior estimates and other variables
produced by \code{forecast()}.
}
\description{
Unnest posterior estimates from a forecast data.frame
}
\examples{
\dontrun{
library(data.table)
options(mc.cores = 4)
dt <- forecast(latest_obs(germany_covid19_delta_obs), max_treedepth = 15)

# unnest posterior predictions
posterior <- unnest_posterior(dt, target = "posterior")
posterior

# unnest forecasts
forecasts <- unnest_posterior(dt, target = "forecast")
forecasts
}
}
