% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{update_voc_label}
\alias{update_voc_label}
\title{Label the Variant of Concern}
\usage{
update_voc_label(posterior, label, target_label = "VOC")
}
\arguments{
\item{posterior}{A list of posterior output as produced by
\code{summarise_posterior()}.}

\item{label}{Character string  indicating the new label to use for the
variant of concern.}

\item{target_label}{A character string defaulting to "VOC". Indicates the
current label for the variant of concern.}
}
\value{
A list of data frames as returned by \code{summarise_posterior()} but
with updated labels.
}
\description{
Assign a custom label to the variant of concern in the
output from \code{summarise_posterior()}.
}
\examples{
\dontrun{
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
p <- summarise_posterior(fit)
p <- update_voc_label(p, "Delta")
p$model[]
}
}
